+++ packages/abi-gen/CHANGELOG.md
@@ -1,6 +1,7 @@
 # CHANGELOG
 
 ## v0.2.0 - _???_
+
 * Added CLI options for explicit specifying location of partials and main template (#346)
 
 ## v0.1.0 - _January 11, 2018_